Refactor DelegationModal component and extract its content to a separate component
- Extracted the content of the DelegationModal component to a new component called DelegationModalContent
- Created a new context called DelegationModalContext to manage the state and functions related to the DelegationModal component
- Moved the useState hooks and related functions from DelegationModal to DelegationModalContext
- Updated the imports and references in DelegationModal to use the new
